# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Separated User Auth(JWT) Example",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/user/login": {
            "post": {
                "tags": [
                    "User Authentication"
                ],
                "summary": "User login",
                "description": "Authenticate user and return JWT tokens",
                "operationId": "f4303455a3fb0154bd62e744e406ba38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token pair response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access": {
                                            "properties": {
                                                "token": {},
                                                "expiredAt": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "refresh": {
                                            "properties": {
                                                "token": {},
                                                "expiredAt": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/user/refresh": {
            "post": {
                "tags": [
                    "User Authentication"
                ],
                "summary": "Refresh JWT token",
                "description": "Refresh user authentication token",
                "operationId": "6c229207ea6e75a2068227ce99221cd8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access": {
                                            "properties": {
                                                "token": {},
                                                "expiredAt": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "refresh": {
                                            "properties": {
                                                "token": {},
                                                "expiredAt": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/user/logout": {
            "post": {
                "tags": [
                    "User Authentication"
                ],
                "summary": "User logout",
                "description": "Invalidate access/refresh token pair for current session",
                "operationId": "05826f6242298a10f44ee156701bed1a",
                "responses": {
                    "200": {
                        "description": "Empty response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {}
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/auth/user/logout/all": {
            "post": {
                "tags": [
                    "User Authentication"
                ],
                "summary": "Logout from all devices",
                "description": "Invalidate access/refresh token pairs for all user sessions",
                "operationId": "59fb7ca40da5fad408e133093ee68a19",
                "responses": {
                    "200": {
                        "description": "Empty response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {}
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/auth/user/profile": {
            "get": {
                "tags": [
                    "User Authentication"
                ],
                "summary": "Get user profile",
                "description": "Returns authenticated user's profile information",
                "operationId": "42d6738579c102d21d5544f4c6a1b807",
                "responses": {
                    "200": {
                        "description": "User profile information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "user@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "JWT": {
                "type": "http",
                "name": "JWT",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}